refactor(#217): move TTS speakers to a file
